be/requirements.txt
fastapi
uvicorn
boto3
pydantic_settings
python-dotenv
sqlalchemy
psycopg2-binary